Use descriptor set generation from ProtoData
Also:
 * Rename `JavaProtocConfigurationPlugin` to `EnableGrpcPlugin`.
 * Introduce utility for forcing Gradle project evaluation.
